python-data-engineer/week9/mini_project/requirements.txt
# Core dependencies
fastapi==0.68.1
uvicorn==0.15.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
pydantic==1.8.2

# Database
psycopg2-binary==2.9.1
pymongo==3.12.0
redis==3.5.3
SQLAlchemy==1.4.23

# Data processing
pandas==1.3.3
numpy==1.21.2
apache-kafka==2.8.0
confluent-kafka==1.7.0

# Visualization
plotly==5.3.1
dash==2.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0

# Testing
pytest==6.2.5
pytest-asyncio==0.15.1
httpx==0.19.0

# Development
black==21.9b0
mypy==0.910
isort==5.9.3

# Monitoring
prometheus-client==0.11.0 